# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test App",
        "version": "1.0.0"
    },
    "paths": {
        "/api/list": {
            "get": {
                "operationId": "b594e98176861eb7cd1a8f75b3bb2b26",
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                }
            }
        },
        "/api/add": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Users",
                "description": "Users here",
                "operationId": "add-Users",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "name",
                                    "email",
                                    "phone",
                                    "education_id",
                                    "gender",
                                    "hobbies",
                                    "experience",
                                    "message"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "file"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "integer"
                                    },
                                    "education_id": {
                                        "type": "integer"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "hobbies": {
                                        "type": "string"
                                    },
                                    "experience": {
                                        "type": "longtext"
                                    },
                                    "message": {
                                        "type": "longtext"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "comment has been created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "comment has been created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete Users",
                "description": "Delete Schedule",
                "operationId": "delete users",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of users",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    }
}